Introduce a user model for the Sequel gem (#56)
We keep track of last_login to effectively identify inactive users.
This used to be done by the logging API but will now be done by this
API.

This makes it easier to split the userdetails table out of the main
database and be owned by this API.

The definition of "last login" will change as a result of this, and will
effectively be "last login attempt" as we will populate this field for
both access accept and access reject.  I don't see this affecting GDPR
cleanup tasks at all.