Change ECS service name in staging
We are in the process of making any references to the authentication api
infrastructure consistent, and we are testing this change in Staging. At
present we are using the name "authentication-api" and "authorization-api" to
refer to the same things, which is confusing. This change is part of rectifying
that, more information is available in this card:
https://technologyprogramme.atlassian.net/jira/software/projects/GW/boards/251?selectedIssue=GW-204